pyproject.toml
[tool.poetry]
name = "pirate"
version = "0.1.1a1"
description = "Pirate is a Python package that aims to be a unified framework for all things information retrieval. Currently, it serves as a go-to library for negative mining."
authors = ["Herumb Shandilya <herumbshandilya123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
polars = "^0.20.16"
mkdocs-material = {extras = ["imaging"], version = "^9.5.14"}
mkdocstrings = {extras = ["crystal", "python"], version = "^0.24.1"}
pytest = "^8.1.1"
rank-bm25 = "^0.2.2"
sentence-transformers = "^2.6.1"
pydantic = "^2.7.1"
loguru = "^0.7.2"
ruff = "^0.4.4"
pyright = "^1.1.362"
dspy-ai = "^2.4.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
